set upstrim
