+++ index.html
@@ -252,7 +252,7 @@
     </div>
   </template>
 
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 
 </html>
